refactor(crypto): Split `types::events::room::encrypted` to modules.
