+++ crates/matrix-sdk-crypto/src/types/events/room/mod.rs
@@ -23,6 +23,9 @@ use serde_json::Value;
 use super::EventType;
 
 pub mod encrypted;
+pub mod room;
+pub mod scheme;
+pub mod to_device;
 
 /// Generic room event with a known type and content.
 #[derive(Debug, Deserialize)]
